upgrade to darwinkit

+++ client/cmd/bisonw-desktop/app.go
@@ -430,7 +430,7 @@ func runWebview(url string) {
 		useIconBytes(w, Icon) // useIcon(w, "src/dexc.png")
 	}
 
-	width, height := limitedWindowWidthAndHeight(int(C.display_width()), int(C.display_height()))
+	width, height := limitedWindowWidthAndHeight(float64(C.display_width()), float64(C.display_height()))
 
 	w.SetSize(width, height, webview.HintNone)
 	w.Navigate(url)
